REFACTORING OF INPUT InputGenerator.java, Replayer.java, and Fuzz.java.
This is a refactoring commit that moves InputGenerator to work with
keyEvents and a single call method. This means that if another part of
the application has added their own bindings, these will be accessible
by the inputGenerator.

We still need the input generator to allow locking of keyEvents from the
keyboard. Just means we can still call movement and other things & have
the replayers keys works.