Refactoring & App Feature: key locking completion.
Keylocking is now fully inplace and the replayer should now set it's
bindings only when they are needed.

Changes made to Replayer.java were needed so I have access to set the
keybindings. A component should only know what it's bindings are, not
where to set them. Where to set them is App's responsibility.